sqli_scanner/requirements.txt
httpx
requests
beautifulsoup4
lxml
PyYAML
rich
colorama
typer
Jinja2
